Create sudoku.hpp and turn struct into class
